@@ -2,11 +2,11 @@ name: cloud_firestore_example
 description: Demonstrates how to use the firestore plugin.
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
-  cloud_firestore: ^5.0.0
-  firebase_core: ^3.0.0
+  cloud_firestore: ^5.1.0
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
   http: ^1.0.0
