@@ -4,7 +4,7 @@ description:
   live synchronization and offline support on Android and iOS.
 homepage: https://firebase.google.com/docs/firestore
 repository: https://github.com/firebase/flutterfire/tree/master/packages/cloud_firestore/cloud_firestore
-version: 5.0.0
+version: 5.1.0
 topics:
   - firebase
   - firestore
@@ -16,15 +16,15 @@ false_secrets:
   - dartpad/**
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  cloud_firestore_platform_interface: ^6.2.6
-  cloud_firestore_web: ^4.0.0
+  cloud_firestore_platform_interface: ^6.2.9
+  cloud_firestore_web: ^4.0.3
   collection: ^1.0.0
-  firebase_core: ^3.0.0
-  firebase_core_platform_interface: ^5.0.0
+  firebase_core: ^3.2.0
+  firebase_core_platform_interface: ^5.1.0
   flutter:
     sdk: flutter
   meta: ^1.8.0
