@@ -1,24 +1,24 @@
 name: cloud_firestore_platform_interface
 description: A common platform interface for the cloud_firestore plugin.
-version: 6.2.6
+version: 6.2.9
 homepage: https://github.com/firebase/flutterfire/tree/master/packages/cloud_firestore/cloud_firestore_platform_interface
 repository: https://github.com/firebase/flutterfire/tree/master/packages/cloud_firestore/cloud_firestore_platform_interface
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  _flutterfire_internals: ^1.3.36
+  _flutterfire_internals: ^1.3.39
   collection: ^1.15.0
-  firebase_core: ^3.0.0
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
   meta: ^1.8.0
   plugin_platform_interface: ^2.1.3
 
 dev_dependencies:
-  firebase_core_platform_interface: ^5.0.0
+  firebase_core_platform_interface: ^5.1.0
   flutter_test:
     sdk: flutter
   mockito: ^5.0.0
