@@ -2,12 +2,12 @@ name: cloud_functions_example
 description: Demonstrates how to use the cloud_functions plugin.
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  cloud_functions: ^5.0.0
-  firebase_core: ^3.0.0
+  cloud_functions: ^5.0.3
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
 
