@@ -1,6 +1,6 @@
 name: cloud_functions
 description: A Flutter plugin allowing you to use Firebase Cloud Functions.
-version: 5.0.0
+version: 5.0.3
 homepage: https://firebase.google.com/docs/functions
 repository: https://github.com/firebase/flutterfire/tree/master/packages/cloud_functions/cloud_functions
 topics:
@@ -13,14 +13,14 @@ false_secrets:
   - example/**
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  cloud_functions_platform_interface: ^5.5.29
-  cloud_functions_web: ^4.9.7
-  firebase_core: ^3.0.0
-  firebase_core_platform_interface: ^5.0.0
+  cloud_functions_platform_interface: ^5.5.32
+  cloud_functions_web: ^4.9.10
+  firebase_core: ^3.2.0
+  firebase_core_platform_interface: ^5.1.0
   flutter:
     sdk: flutter
 
