@@ -5,21 +5,21 @@ repository: https://github.com/firebase/flutterfire/tree/master/packages/cloud_f
 
 # NOTE: We strongly prefer non-breaking changes, even at the expense of a
 # less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
-version: 5.5.29
+version: 5.5.32
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  firebase_core: ^3.0.0
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
   meta: ^1.8.0
   plugin_platform_interface: ^2.1.3
 
 dev_dependencies:
-  firebase_core_platform_interface: ^5.0.0
+  firebase_core_platform_interface: ^5.1.0
   flutter_test:
     sdk: flutter
   mockito: ^5.0.0
