@@ -2,12 +2,12 @@ name: firebase_analytics_example
 description: Demonstrates how to use the firebase_analytics plugin.
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  firebase_analytics: ^11.0.0
-  firebase_core: ^3.0.0
+  firebase_analytics: ^11.2.0
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
 
