@@ -4,7 +4,7 @@ description:
   solution that provides insight on app usage and user engagement on Android and iOS.
 homepage: https://firebase.google.com/docs/analytics
 repository: https://github.com/firebase/flutterfire/tree/master/packages/firebase_analytics/firebase_analytics
-version: 11.0.0
+version: 11.2.0
 topics:
   - firebase
   - analytics
@@ -15,14 +15,14 @@ false_secrets:
   - example/**
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  firebase_analytics_platform_interface: ^4.0.0
-  firebase_analytics_web: ^0.5.7+8
-  firebase_core: ^3.0.0
-  firebase_core_platform_interface: ^5.0.0
+  firebase_analytics_platform_interface: ^4.2.0
+  firebase_analytics_web: ^0.5.9
+  firebase_core: ^3.2.0
+  firebase_core_platform_interface: ^5.1.0
   flutter:
     sdk: flutter
 
