@@ -2,21 +2,21 @@ name: firebase_analytics_platform_interface
 description: A common platform interface for the firebase_analytics plugin.
 homepage: https://github.com/firebase/flutterfire/tree/master/packages/firebase_analytics/firebase_analytics_platform_interface
 repository: https://github.com/firebase/flutterfire/tree/master/packages/firebase_analytics/firebase_analytics_platform_interface
-version: 4.0.0
+version: 4.2.0
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   flutter: '>=3.3.0'
 
 dependencies:
-  _flutterfire_internals: ^1.3.36
-  firebase_core: ^3.0.0
+  _flutterfire_internals: ^1.3.39
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
   meta: ^1.8.0
   plugin_platform_interface: ^2.1.3
 
 dev_dependencies:
-  firebase_core_platform_interface: ^5.0.0
+  firebase_core_platform_interface: ^5.1.0
   flutter_test:
     sdk: flutter
