@@ -6,12 +6,12 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 dependencies:
-  cloud_firestore: ^5.0.0
-  firebase_app_check: ^0.3.0
-  firebase_core: ^3.0.0
+  cloud_firestore: ^5.1.0
+  firebase_app_check: ^0.3.0+3
+  firebase_core: ^3.2.0
   flutter:
     sdk: flutter
 
