@@ -100,11 +100,24 @@ class Firestore extends JsObjectWrapper<firestore_interop.FirestoreJsImpl> {
         firestore_interop.enableIndexedDbPersistence(jsObject).toDart;
   }
 
-  String get _snapshotInSyncWindowsKey =>
-      'flutterfire-${app.name}_snapshotInSync';
+// purely for debug mode and tracking listeners to clean up on "hot restart"
+  final Map<String, int> _snapshotInSyncListeners = {};
+  String _snapshotInSyncWindowsKey() {
+    if (kDebugMode) {
+      final key = 'flutterfire-${app.name}_snapshotInSync';
+      if (_snapshotInSyncListeners.containsKey(key)) {
+        _snapshotInSyncListeners[key] = _snapshotInSyncListeners[key]! + 1;
+      } else {
+        _snapshotInSyncListeners[key] = 0;
+      }
+      return '$key-${_snapshotInSyncListeners[key]}';
+    }
+    return 'no-op';
+  }
 
   Stream<void> snapshotsInSync() {
-    unsubscribeWindowsListener(_snapshotInSyncWindowsKey);
+    final snapshotKey = _snapshotInSyncWindowsKey();
+    unsubscribeWindowsListener(snapshotKey);
     late StreamController<void> controller;
     late JSFunction onSnapshotsInSyncUnsubscribe;
     var nextWrapper = ((JSObject? noValue) {
@@ -115,7 +128,7 @@ class Firestore extends JsObjectWrapper<firestore_interop.FirestoreJsImpl> {
       onSnapshotsInSyncUnsubscribe =
           firestore_interop.onSnapshotsInSync(jsObject, nextWrapper);
       setWindowsListener(
-        _snapshotInSyncWindowsKey,
+        snapshotKey,
         onSnapshotsInSyncUnsubscribe,
       );
     }
@@ -123,7 +136,7 @@ class Firestore extends JsObjectWrapper<firestore_interop.FirestoreJsImpl> {
     void stopListen() {
       onSnapshotsInSyncUnsubscribe.callAsFunction();
       controller.close();
-      removeWindowsListener(_snapshotInSyncWindowsKey);
+      removeWindowsListener(snapshotKey);
     }
 
     controller = StreamController<void>.broadcast(
@@ -373,8 +386,20 @@ class DocumentReference
         (result)! as firestore_interop.DocumentSnapshotJsImpl);
   }
 
-  String get _documentSnapshotWindowsKey =>
-      'flutterfire-${firestore.app.name}_${path}_documentSnapshot';
+  // purely for debug mode and tracking listeners to clean up on "hot restart"
+  final Map<String, int> _docListeners = {};
+  String _documentSnapshotWindowsKey() {
+    if (kDebugMode) {
+      final key = 'flutterfire-${firestore.app.name}_${path}_documentSnapshot';
+      if (_docListeners.containsKey(key)) {
+        _docListeners[key] = _docListeners[key]! + 1;
+      } else {
+        _docListeners[key] = 0;
+      }
+      return '$key-${_docListeners[key]}';
+    }
+    return 'no-op';
+  }
 
   /// Attaches a listener for [DocumentSnapshot] events.
   Stream<DocumentSnapshot> onSnapshot({
@@ -391,7 +416,8 @@ class DocumentReference
   StreamController<DocumentSnapshot> _createSnapshotStream([
     firestore_interop.DocumentListenOptions? options,
   ]) {
-    unsubscribeWindowsListener(_documentSnapshotWindowsKey);
+    final documentKey = _documentSnapshotWindowsKey();
+    unsubscribeWindowsListener(documentKey);
     late JSFunction onSnapshotUnsubscribe;
     // ignore: close_sinks, the controller is returned
     late StreamController<DocumentSnapshot> controller;
@@ -408,12 +434,12 @@ class DocumentReference
               jsObject as JSObject, options as JSAny, nextWrapper, errorWrapper)
           : firestore_interop.onSnapshot(
               jsObject as JSObject, nextWrapper, errorWrapper);
-      setWindowsListener(_documentSnapshotWindowsKey, onSnapshotUnsubscribe);
+      setWindowsListener(documentKey, onSnapshotUnsubscribe);
     }
 
     void stopListen() {
       onSnapshotUnsubscribe.callAsFunction();
-      removeWindowsListener(_documentSnapshotWindowsKey);
+      removeWindowsListener(documentKey);
     }
 
     return controller = StreamController<DocumentSnapshot>.broadcast(
@@ -486,8 +512,20 @@ class Query<T extends firestore_interop.QueryJsImpl>
   Query limitToLast(num limit) => Query.fromJsObject(firestore_interop.query(
       jsObject, firestore_interop.limitToLast(limit.toJS)));
 
-  String _querySnapshotWindowsKey(hashCode) =>
-      'flutterfire-${firestore.app.name}_${hashCode}_querySnapshot';
+  // purely for debug mode and tracking listeners to clean up on "hot restart"
+  final Map<String, int> _snapshotListeners = {};
+  String _querySnapshotWindowsKey(hashCode) {
+    if (kDebugMode) {
+      final key = 'flutterfire-${firestore.app.name}_${hashCode}_querySnapshot';
+      if (_snapshotListeners.containsKey(key)) {
+        _snapshotListeners[key] = _snapshotListeners[key]! + 1;
+      } else {
+        _snapshotListeners[key] = 0;
+      }
+      return '$key-${_snapshotListeners[key]}';
+    }
+    return 'no-op';
+  }
 
   Stream<QuerySnapshot> onSnapshot(
           {bool includeMetadataChanges = false,
@@ -505,7 +543,8 @@ class Query<T extends firestore_interop.QueryJsImpl>
     firestore_interop.DocumentListenOptions options,
     int hashCode,
   ) {
-    unsubscribeWindowsListener(_querySnapshotWindowsKey(hashCode));
+    final snapshotKey = _querySnapshotWindowsKey(hashCode);
+    unsubscribeWindowsListener(snapshotKey);
     late JSFunction onSnapshotUnsubscribe;
     // ignore: close_sinks, the controller is returned
     late StreamController<QuerySnapshot> controller;
@@ -519,12 +558,14 @@ class Query<T extends firestore_interop.QueryJsImpl>
       onSnapshotUnsubscribe = firestore_interop.onSnapshot(
           jsObject as JSObject, options as JSObject, nextWrapper, errorWrapper);
       setWindowsListener(
-          _querySnapshotWindowsKey(hashCode), onSnapshotUnsubscribe);
+        snapshotKey,
+        onSnapshotUnsubscribe,
+      );
     }
 
     void stopListen() {
       onSnapshotUnsubscribe.callAsFunction();
-      removeWindowsListener(_querySnapshotWindowsKey(hashCode));
+      removeWindowsListener(snapshotKey);
     }
 
     return controller = StreamController<QuerySnapshot>.broadcast(
