@@ -34,7 +34,10 @@ class MethodChannelFirebaseAnalytics extends FirebaseAnalyticsPlatform {
       MethodChannel('plugins.flutter.io/firebase_analytics');
 
   @override
-  FirebaseAnalyticsPlatform delegateFor({required FirebaseApp app}) {
+  FirebaseAnalyticsPlatform delegateFor({
+    required FirebaseApp app,
+    Map<String, dynamic>? webOptions,
+  }) {
     return MethodChannelFirebaseAnalytics(app: app);
   }
 
@@ -214,6 +217,8 @@ class MethodChannelFirebaseAnalytics extends FirebaseAnalyticsPlatform {
   Future<void> initiateOnDeviceConversionMeasurement({
     String? emailAddress,
     String? phoneNumber,
+    String? hashedEmailAddress,
+    String? hashedPhoneNumber,
   }) {
     try {
       return channel.invokeMethod<void>(
@@ -221,6 +226,8 @@ class MethodChannelFirebaseAnalytics extends FirebaseAnalyticsPlatform {
         <String, String?>{
           'emailAddress': emailAddress,
           'phoneNumber': phoneNumber,
+          'hashedEmailAddress': hashedEmailAddress,
+          'hashedPhoneNumber': hashedPhoneNumber,
         },
       );
     } catch (e, s) {
