@@ -8,9 +8,9 @@
 #import <Flutter/Flutter.h>
 #endif
 
-#import <FLTAppCheckProviderFactory.h>
 #import <Foundation/Foundation.h>
 #import <firebase_core/FLTFirebasePlugin.h>
+#import "FLTAppCheckProviderFactory.h"
 
 @interface FLTFirebaseAppCheckPlugin : FLTFirebasePlugin <FlutterPlugin, FLTFirebasePlugin>
 @end
