@@ -52,6 +52,12 @@ NSString *const kArgumentToken = @"token";
 NSString *const kArgumentVerificationId = @"verificationId";
 NSString *const kArgumentSmsCode = @"smsCode";
 NSString *const kArgumentActionCodeSettings = @"actionCodeSettings";
+NSString *const kArgumentFamilyName = @"familyName";
+NSString *const kArgumentGivenName = @"givenName";
+NSString *const kArgumentMiddleName = @"middleName";
+NSString *const kArgumentNickname = @"nickname";
+NSString *const kArgumentNamePrefix = @"namePrefix";
+NSString *const kArgumentNameSuffix = @"nameSuffix";
 
 // MultiFactor
 NSString *const kArgumentMultiFactorHints = @"multiFactorHints";
@@ -187,13 +193,14 @@ static NSMutableDictionary<NSNumber *, FIRAuthCredential *> *credentialsMap;
     additionalData[kArgumentEmail] = [error userInfo][FIRAuthErrorUserInfoEmailKey];
   }
   // We want to store the credential if present for future sign in if the exception contains a
-  // credential
-  [FLTFirebaseAuthPlugin storeAuthCredentialIfPresent:error];
+  // credential, we pass a token back to Flutter to allow retreival of the credential.
+  NSNumber *token = [FLTFirebaseAuthPlugin storeAuthCredentialIfPresent:error];
 
   // additionalData.authCredential
   if ([error userInfo][FIRAuthErrorUserInfoUpdatedCredentialKey] != nil) {
     FIRAuthCredential *authCredential = [error userInfo][FIRAuthErrorUserInfoUpdatedCredentialKey];
-    additionalData[@"authCredential"] = [PigeonParser getPigeonAuthCredential:authCredential];
+    additionalData[@"authCredential"] = [PigeonParser getPigeonAuthCredential:authCredential
+                                                                        token:token];
   }
 
   // Manual message overrides to ensure messages/codes matches other platforms.
@@ -635,14 +642,16 @@ static void handleAppleAuthResult(FLTFirebaseAuthPlugin *object, AuthPigeonFireb
 
 #pragma mark - Utilities
 
-+ (void)storeAuthCredentialIfPresent:(NSError *)error {
++ (NSNumber *_Nullable)storeAuthCredentialIfPresent:(NSError *)error {
   if ([error userInfo][FIRAuthErrorUserInfoUpdatedCredentialKey] != nil) {
     FIRAuthCredential *authCredential = [error userInfo][FIRAuthErrorUserInfoUpdatedCredentialKey];
     // We temporarily store the non-serializable credential so the
     // Dart API can consume these at a later time.
     NSNumber *authCredentialHash = @([authCredential hash]);
     credentialsMap[authCredentialHash] = authCredential;
+    return authCredentialHash;
   }
+  return nil;
 }
 
 - (FIRAuth *_Nullable)getFIRAuthFromAppNameFromPigeon:(AuthPigeonFirebaseApp *)pigeonApp {
@@ -754,7 +763,31 @@ static void handleAppleAuthResult(FLTFirebaseAuthPlugin *object, AuthPigeonFireb
     return;
 #endif
   }
+  // Apple Auth
+  if ([signInMethod isEqualToString:kSignInMethodApple]) {
+    if (idToken && rawNonce) {
+      // Credential with idToken, rawNonce and fullName
+      NSPersonNameComponents *fullName = [[NSPersonNameComponents alloc] init];
+      fullName.givenName =
+          arguments[kArgumentGivenName] == [NSNull null] ? nil : arguments[kArgumentGivenName];
+      fullName.familyName =
+          arguments[kArgumentFamilyName] == [NSNull null] ? nil : arguments[kArgumentFamilyName];
+      fullName.nickname =
+          arguments[kArgumentNickname] == [NSNull null] ? nil : arguments[kArgumentNickname];
+      fullName.namePrefix =
+          arguments[kArgumentNamePrefix] == [NSNull null] ? nil : arguments[kArgumentNamePrefix];
+      fullName.nameSuffix =
+          arguments[kArgumentNameSuffix] == [NSNull null] ? nil : arguments[kArgumentNameSuffix];
+      fullName.middleName =
+          arguments[kArgumentMiddleName] == [NSNull null] ? nil : arguments[kArgumentMiddleName];
 
+      completion([FIROAuthProvider appleCredentialWithIDToken:idToken
+                                                     rawNonce:rawNonce
+                                                     fullName:fullName],
+                 nil);
+      return;
+    }
+  }
   // OAuth
   if ([signInMethod isEqualToString:kSignInMethodOAuth]) {
     NSString *providerId = arguments[kArgumentProviderId];
