@@ -46,6 +46,19 @@ class AppleAuthProvider extends AuthProvider {
     );
   }
 
+  /// Create a new [AppleAuthCredential] from a provided [idToken], [rawNonce] and [appleFullPersonName];
+  static OAuthCredential credentialWithIDToken(
+    String idToken,
+    String rawNonce,
+    AppleFullPersonName appleFullPersonName,
+  ) {
+    return AppleAuthCredential._credentialWithIDToken(
+      idToken,
+      rawNonce,
+      appleFullPersonName,
+    );
+  }
+
   /// This corresponds to the sign-in method identifier.
   static String get APPLE_SIGN_IN_METHOD {
     return _kProviderId;
@@ -89,13 +102,52 @@ class AppleAuthProvider extends AuthProvider {
 /// [AppleAuthProvider.credential].
 class AppleAuthCredential extends OAuthCredential {
   AppleAuthCredential._({
-    required String accessToken,
+    String? accessToken,
+    String? rawNonce,
+    String? idToken,
+    AppleFullPersonName? appleFullPersonName,
   }) : super(
-            providerId: _kProviderId,
-            signInMethod: _kProviderId,
-            accessToken: accessToken);
+          providerId: _kProviderId,
+          signInMethod: _kProviderId,
+          accessToken: accessToken,
+          appleFullPersonName: appleFullPersonName,
+          rawNonce: rawNonce,
+          idToken: idToken,
+        );
 
   factory AppleAuthCredential._credential(String accessToken) {
-    return AppleAuthCredential._(accessToken: accessToken);
+    return AppleAuthCredential._(
+      accessToken: accessToken,
+    );
   }
+
+  factory AppleAuthCredential._credentialWithIDToken(
+    String idToken,
+    String rawNonce,
+    AppleFullPersonName appleFullPersonName,
+  ) {
+    return AppleAuthCredential._(
+      idToken: idToken,
+      rawNonce: rawNonce,
+      appleFullPersonName: appleFullPersonName,
+    );
+  }
+}
+
+class AppleFullPersonName {
+  AppleFullPersonName({
+    this.givenName,
+    this.familyName,
+    this.middleName,
+    this.nickname,
+    this.namePrefix,
+    this.nameSuffix,
+  });
+
+  final String? givenName;
+  final String? familyName;
+  final String? middleName;
+  final String? nickname;
+  final String? namePrefix;
+  final String? nameSuffix;
 }
