@@ -2,16 +2,15 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library firebase_database_web;
-
 import 'dart:async';
 import 'dart:js_interop';
-
+import 'package:collection/collection.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
 import 'package:firebase_core_web/firebase_core_web_interop.dart'
     as core_interop;
 import 'package:firebase_database_platform_interface/firebase_database_platform_interface.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 import 'src/interop/database.dart' as database_interop;
@@ -48,8 +47,7 @@ class FirebaseDatabaseWeb extends DatabasePlatform {
 
   /// Builds an instance of [DatabaseWeb] with an optional [FirebaseApp] instance
   /// If [app] is null then the created instance will use the default [FirebaseApp]
-  FirebaseDatabaseWeb({FirebaseApp? app, String? databaseURL})
-      : super(app: app, databaseURL: databaseURL);
+  FirebaseDatabaseWeb({super.app, super.databaseURL});
 
   @override
   DatabasePlatform delegateFor(
