@@ -51,8 +51,7 @@ class Database
   static Database getInstance(database_interop.DatabaseJsImpl jsObject) =>
       _expando[jsObject] ??= Database._fromJsObject(jsObject);
 
-  Database._fromJsObject(database_interop.DatabaseJsImpl jsObject)
-      : super.fromJsObject(jsObject);
+  Database._fromJsObject(super.jsObject) : super.fromJsObject();
 
   /// Disconnects from the server, all database operations will be
   /// completed offline.
@@ -103,7 +102,7 @@ class DatabaseReference<T extends database_interop.ReferenceJsImpl>
   ) =>
       _expando[jsObject] ??= DatabaseReference._fromJsObject(jsObject);
 
-  DatabaseReference._fromJsObject(T jsObject) : super.fromJsObject(jsObject);
+  DatabaseReference._fromJsObject(super.jsObject) : super.fromJsObject();
 
   /// Returns child DatabaseReference from provided relative [path].
   DatabaseReference child(String path) => DatabaseReference.getInstance(
@@ -251,40 +250,67 @@ class Query<T extends database_interop.QueryJsImpl> extends JsObjectWrapper<T> {
   /// DatabaseReference to the Query's location.
   DatabaseReference get ref => DatabaseReference.getInstance(jsObject.ref);
 
-  late final Stream<QueryEvent> _onValue = _createStream('value');
+  Stream<QueryEvent> _onValue(String appName, String hashCode) => _createStream(
+        'value',
+        appName,
+        hashCode,
+      );
 
   /// Stream for a value event. Event is triggered once with the initial
   /// data stored at location, and then again each time the data changes.
-  Stream<QueryEvent> get onValue => _onValue;
+  Stream<QueryEvent> onValue(String appName, String hashCode) =>
+      _onValue(appName, hashCode);
 
-  late final Stream<QueryEvent> _onChildAdded = _createStream('child_added');
+  Stream<QueryEvent> _onChildAdded(String appName, String hashCode) =>
+      _createStream(
+        'child_added',
+        appName,
+        hashCode,
+      );
 
   /// Stream for a child_added event. Event is triggered once for each
   /// initial child at location, and then again every time a new child is added.
-  Stream<QueryEvent> get onChildAdded => _onChildAdded;
+  Stream<QueryEvent> onChildAdded(String appName, String hashCode) =>
+      _onChildAdded(appName, hashCode);
 
-  late final Stream<QueryEvent> _onChildRemoved =
-      _createStream('child_removed');
+  Stream<QueryEvent> _onChildRemoved(String appName, String hashCode) =>
+      _createStream(
+        'child_removed',
+        appName,
+        hashCode,
+      );
 
   /// Stream for a child_removed event. Event is triggered once every time
   /// a child is removed.
-  Stream<QueryEvent> get onChildRemoved => _onChildRemoved;
+  Stream<QueryEvent> onChildRemoved(String appName, String hashCode) =>
+      _onChildRemoved(appName, hashCode);
 
-  late final Stream<QueryEvent> _onChildChanged =
-      _createStream('child_changed');
+  Stream<QueryEvent> _onChildChanged(String appName, String hashCode) =>
+      _createStream(
+        'child_changed',
+        appName,
+        hashCode,
+      );
 
   /// Stream for a child_changed event. Event is triggered when the data
   /// stored in a child (or any of its descendants) changes.
   /// Single child_changed event may represent multiple changes to the child.
-  Stream<QueryEvent> get onChildChanged => _onChildChanged;
-  late final Stream<QueryEvent> _onChildMoved = _createStream('child_moved');
+  Stream<QueryEvent> onChildChanged(String appName, String hashCode) =>
+      _onChildChanged(appName, hashCode);
+  Stream<QueryEvent> _onChildMoved(String appName, String hashCode) =>
+      _createStream(
+        'child_moved',
+        appName,
+        hashCode,
+      );
 
   /// Stream for a child_moved event. Event is triggered when a child's priority
   /// changes such that its position relative to its siblings changes.
-  Stream<QueryEvent> get onChildMoved => _onChildMoved;
+  Stream<QueryEvent> onChildMoved(String appName, String hashCode) =>
+      _onChildMoved(appName, hashCode);
 
   /// Creates a new Query from a [jsObject].
-  Query.fromJsObject(T jsObject) : super.fromJsObject(jsObject);
+  Query.fromJsObject(super.jsObject) : super.fromJsObject();
 
   /// Gets the most up-to-date result for this query.
   Future<DataSnapshot> get() async {
@@ -377,66 +403,86 @@ class Query<T extends database_interop.QueryJsImpl> extends JsObjectWrapper<T> {
     );
   }
 
-  Stream<QueryEvent> _createStream(String eventType) {
-    late StreamController<QueryEvent> streamController;
+  String _streamWindowsKey(String appName, String eventType, String hashCode) =>
+      'flutterfire-${appName}_${eventType}_${hashCode}_snapshot';
 
+  Stream<QueryEvent> _createStream(
+    String eventType,
+    String appName,
+    String hashCode,
+  ) {
+    late StreamController<QueryEvent> streamController;
+    unsubscribeWindowsListener(_streamWindowsKey(appName, eventType, hashCode));
     final callbackWrap = ((
       database_interop.DataSnapshotJsImpl data, [
-      String? string,
+      String? prevChild,
     ]) {
-      streamController.add(QueryEvent(DataSnapshot.getInstance(data), string));
+      streamController
+          .add(QueryEvent(DataSnapshot.getInstance(data), prevChild));
     });
 
     final void Function(JSObject) cancelCallbackWrap = ((JSObject error) {
       streamController.addError(convertFirebaseDatabaseException(error));
-      streamController.close();
     });
+
+    late JSFunction onUnsubscribe;
 
     void startListen() {
       if (eventType == 'child_added') {
-        database_interop.onChildAdded(
+        onUnsubscribe = database_interop.onChildAdded(
           jsObject,
           callbackWrap.toJS,
           cancelCallbackWrap.toJS,
         );
       }
       if (eventType == 'value') {
-        database_interop.onValue(
+        onUnsubscribe = database_interop.onValue(
           jsObject,
           callbackWrap.toJS,
           cancelCallbackWrap.toJS,
         );
       }
       if (eventType == 'child_removed') {
-        database_interop.onChildRemoved(
+        onUnsubscribe = database_interop.onChildRemoved(
           jsObject,
           callbackWrap.toJS,
           cancelCallbackWrap.toJS,
         );
       }
       if (eventType == 'child_changed') {
-        database_interop.onChildChanged(
+        onUnsubscribe = database_interop.onChildChanged(
           jsObject,
           callbackWrap.toJS,
           cancelCallbackWrap.toJS,
         );
       }
       if (eventType == 'child_moved') {
-        database_interop.onChildMoved(
+        onUnsubscribe = database_interop.onChildMoved(
           jsObject,
           callbackWrap.toJS,
           cancelCallbackWrap.toJS,
         );
       }
+      setWindowsListener(
+        _streamWindowsKey(appName, eventType, hashCode),
+        onUnsubscribe,
+      );
     }
 
     void stopListen() {
-      database_interop.off(jsObject, eventType.toJS, callbackWrap.toJS);
+      onUnsubscribe.callAsFunction();
+      streamController.close();
+      removeWindowsListener(_streamWindowsKey(
+        appName,
+        eventType,
+        hashCode,
+      ));
     }
 
     streamController = StreamController<QueryEvent>.broadcast(
       onListen: startListen,
       onCancel: stopListen,
+      sync: true,
     );
     return streamController.stream;
   }
@@ -447,8 +493,8 @@ class Query<T extends database_interop.QueryJsImpl> extends JsObjectWrapper<T> {
 
     database_interop.onValue(
       jsObject,
-      ((database_interop.DataSnapshotJsImpl snapshot, [String? string]) {
-        c.complete(QueryEvent(DataSnapshot.getInstance(snapshot), string));
+      ((database_interop.DataSnapshotJsImpl snapshot, [String? prevChild]) {
+        c.complete(QueryEvent(DataSnapshot.getInstance(snapshot), prevChild));
       }).toJS,
       ((JSAny error) {
         c.completeError(convertFirebaseDatabaseException(error));
@@ -525,9 +571,7 @@ class TransactionResult
   ) =>
       _expando[jsObject] ??= TransactionResult._fromJsObject(jsObject);
 
-  TransactionResult._fromJsObject(
-      database_interop.TransactionResultJsImpl jsObject)
-      : super.fromJsObject(jsObject);
+  TransactionResult._fromJsObject(super.jsObject) : super.fromJsObject();
 
   bool get committed => jsObject.committed.toDart;
 
@@ -555,8 +599,7 @@ class DataSnapshot
   ) =>
       _expando[jsObject] ??= DataSnapshot._fromJsObject(jsObject);
 
-  DataSnapshot._fromJsObject(database_interop.DataSnapshotJsImpl jsObject)
-      : super.fromJsObject(jsObject);
+  DataSnapshot._fromJsObject(super.jsObject) : super.fromJsObject();
 
   /// Returns DataSnapshot for the location at the specified relative [path].
   DataSnapshot child(String path) =>
@@ -598,8 +641,7 @@ class DataSnapshot
 /// See: <https://firebase.google.com/docs/reference/js/firebase.database.OnDisconnect>.
 class OnDisconnect
     extends JsObjectWrapper<database_interop.OnDisconnectJsImpl> {
-  OnDisconnect.fromJsObject(database_interop.OnDisconnectJsImpl jsObject)
-      : super.fromJsObject(jsObject);
+  OnDisconnect.fromJsObject(super.jsObject) : super.fromJsObject();
 
   /// Cancels all previously queued onDisconnect() events for actual location
   /// and all children.
@@ -644,8 +686,8 @@ class ThenableReference
 
   /// Creates a new ThenableReference from a [jsObject].
   ThenableReference.fromJsObject(
-    database_interop.ThenableReferenceJsImpl jsObject,
-  ) : super._fromJsObject(jsObject);
+    super.jsObject,
+  ) : super._fromJsObject();
 
   /// A Future property.
   Future<DatabaseReference> get future => _future;
@@ -670,6 +712,5 @@ class Transaction extends JsObjectWrapper<database_interop.TransactionJsImpl> {
       );
 
   /// Creates a new Transaction from a [jsObject].
-  Transaction.fromJsObject(database_interop.TransactionJsImpl jsObject)
-      : super.fromJsObject(jsObject);
+  Transaction.fromJsObject(super.jsObject) : super.fromJsObject();
 }
