@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of firebase_database_web;
+part of '../firebase_database_web.dart';
 
 /// An implementation of [QueryPlatform] which proxies calls to js objects
 class QueryWeb extends QueryPlatform {
@@ -64,6 +64,8 @@ class QueryWeb extends QueryPlatform {
     return instance;
   }
 
+  final Map<String, int> _streamHashCodeMap = {};
+
   @override
   String get path {
     final refPath = Uri.parse(_queryDelegate.ref.toString()).path;
@@ -89,34 +91,88 @@ class QueryWeb extends QueryPlatform {
     throw UnsupportedError('keepSynced() is not supported on web');
   }
 
+  String _createHashCode(
+    QueryModifiers modifiers,
+    DatabaseEventType eventType,
+    String appName,
+  ) {
+    String hashCode = '0';
+    if (kDebugMode) {
+      hashCode = Object.hashAll([
+        appName,
+        path,
+        ...modifiers
+            .toList()
+            .map((e) => const DeepCollectionEquality().hash(e)),
+        eventType.index,
+      ]).toString();
+      // Need to track as the same properties to create hash could be used multiple times
+      if (_streamHashCodeMap.containsKey(hashCode)) {
+        int count = _streamHashCodeMap[hashCode] ?? 0;
+        final updatedCount = count + 1;
+        _streamHashCodeMap[hashCode] = updatedCount;
+        hashCode = '$hashCode-$updatedCount';
+      } else {
+        // initial stream
+        _streamHashCodeMap[hashCode] = 0;
+        hashCode = '$hashCode-0';
+      }
+    }
+    return hashCode;
+  }
+
   @override
   Stream<DatabaseEventPlatform> observe(
       QueryModifiers modifiers, DatabaseEventType eventType) {
     database_interop.Query instance = _getQueryDelegateInstance(modifiers);
+    final appName =
+        _database.app != null ? _database.app!.name : Firebase.app().name;
+
+    // Purely for unsubscribing purposes in debug mode on "hot restart"
+    // if not running in debug mode, hashCode won't be used
+    String hashCode = _createHashCode(modifiers, eventType, appName);
 
     switch (eventType) {
       case DatabaseEventType.childAdded:
         return _webStreamToPlatformStream(
           eventType,
-          instance.onChildAdded,
+          instance.onChildAdded(
+            appName,
+            hashCode,
+          ),
         );
       case DatabaseEventType.childChanged:
         return _webStreamToPlatformStream(
           eventType,
-          instance.onChildChanged,
+          instance.onChildChanged(
+            appName,
+            hashCode,
+          ),
         );
       case DatabaseEventType.childMoved:
         return _webStreamToPlatformStream(
           eventType,
-          instance.onChildMoved,
+          instance.onChildMoved(
+            appName,
+            hashCode,
+          ),
         );
       case DatabaseEventType.childRemoved:
         return _webStreamToPlatformStream(
           eventType,
-          instance.onChildRemoved,
+          instance.onChildRemoved(
+            appName,
+            hashCode,
+          ),
         );
       case DatabaseEventType.value:
-        return _webStreamToPlatformStream(eventType, instance.onValue);
+        return _webStreamToPlatformStream(
+          eventType,
+          instance.onValue(
+            appName,
+            hashCode,
+          ),
+        );
       default:
         throw Exception("Invalid event type: $eventType");
     }
