@@ -28,6 +28,8 @@ class TaskWeb extends TaskPlatform {
 
   final storage_interop.UploadTask _task;
 
+  Stream<TaskSnapshotPlatform>? _stream;
+
   /// Returns a [Stream] of [TaskSnapshot] events.
   ///
   /// If the task is canceled or fails, the stream will send an error event.
@@ -37,7 +39,7 @@ class TaskWeb extends TaskPlatform {
   /// wait for the stream to complete via [onComplete].
   @override
   Stream<TaskSnapshotPlatform> get snapshotEvents {
-    return guard(() {
+    _stream ??= guard(() {
       // The mobile version of the plugin pushes a "success" snapshot to the
       // onStateChanged stream, but the Firebase JS SDK does *not*.
       // We use a StreamGroup + Future.asStream to simulate that feature:
@@ -46,8 +48,13 @@ class TaskWeb extends TaskPlatform {
 
       // This stream converts the UploadTask Snapshots from JS to the plugins'
       // It can also throw a FirebaseError internally, so we handle it.
-      final onStateChangedStream =
-          _task.onStateChanged.map<TaskSnapshotPlatform>((snapshot) {
+      final onStateChangedStream = _task
+          .onStateChanged(
+        _reference.storage.app.name,
+        _reference.bucket,
+        _reference.fullPath,
+      )
+          .map<TaskSnapshotPlatform>((snapshot) {
         return fbUploadTaskSnapshotToTaskSnapshot(_reference, snapshot);
       });
 
@@ -66,6 +73,8 @@ class TaskWeb extends TaskPlatform {
 
       return group.stream;
     });
+
+    return _stream!;
   }
 
   /// Returns a [Future] once the task has completed.
@@ -101,8 +110,7 @@ class TaskWeb extends TaskPlatform {
     final paused = _task.pause();
     // Wait until the snapshot is paused, then return the value of paused...
     return snapshotEvents
-        .takeWhile((snapshot) => snapshot.state != TaskState.paused)
-        .last
+        .firstWhere((snapshot) => snapshot.state == TaskState.paused)
         .then<bool>((_) => paused);
   }
 
